+++ src/snowflake/snowpark/modin/plugin/compiler/snowflake_query_compiler.py
@@ -3,6 +3,7 @@
 #
 import calendar
 import collections
+import copy
 import functools
 import inspect
 import itertools
@@ -14,7 +15,7 @@ import uuid
 from collections.abc import Hashable, Iterable, Mapping, Sequence
 from datetime import timedelta, tzinfo
 from functools import reduce
-from typing import Any, Callable, List, Literal, Optional, Union, get_args
+from typing import Any, Callable, List, Literal, Optional, TypeVar, Union, get_args
 
 import modin.pandas as pd
 import numpy as np
@@ -411,7 +412,95 @@ _TIMEDELTA_ROLLING_CORR_NOT_SUPPORTED = (
     "ops for Rolling for this dtype timedelta64[ns] are not implemented"
 )
 
+# List of query compiler methods where attrs on the result should always be empty.
+_RESET_ATTRS_METHODS = [
+    "compare",
+    "merge",
+    "value_counts",
+    "dataframe_to_datetime",
+    "series_to_datetime",
+    "to_numeric",
+    "dt_isocalendar",
+    "groupby_all",
+    "groupby_any",
+    "groupby_cumcount",
+    "groupby_cummax",
+    "groupby_cummin",
+    "groupby_cumsum",
+    "groupby_nunique",
+    "groupby_rank",
+    "groupby_size",
+    # expanding and rolling methods also do not propagate; we check them by prefix matching
+    # agg, crosstab, and concat depend on their inputs, and are handled separately
+]
 
+
+T = TypeVar("T", bound=Callable[..., Any])
+
+
+def _propagate_attrs_on_methods(cls):  # type: ignore
+    """
+    Decorator that modifies all methods on the class to copy `_attrs` from `self`
+    to the output of the method, if the output is another query compiler.
+    """
+
+    def propagate_attrs_decorator(method: T) -> T:
+        @functools.wraps(method)
+        def wrap(self, *args, **kwargs):  # type: ignore
+            result = method(self, *args, **kwargs)
+            if isinstance(result, SnowflakeQueryCompiler) and len(self._attrs):
+                result._attrs = copy.deepcopy(self._attrs)
+            return result
+
+        return typing.cast(T, wrap)
+
+    def reset_attrs_decorator(method: T) -> T:
+        @functools.wraps(method)
+        def wrap(self, *args, **kwargs):  # type: ignore
+            result = method(self, *args, **kwargs)
+            if isinstance(result, SnowflakeQueryCompiler) and len(self._attrs):
+                result._attrs = {}
+            return result
+
+        return typing.cast(T, wrap)
+
+    for attr_name, attr_value in cls.__dict__.items():
+        # concat is handled explicitly because it checks all of its arguments
+        # agg is handled explicitly because it sometimes resets and sometimes propagates
+        if attr_name.startswith("_") or attr_name in ["concat", "agg"]:
+            continue
+        if attr_name in _RESET_ATTRS_METHODS or any(
+            attr_name.startswith(prefix) for prefix in ["expanding", "rolling"]
+        ):
+            setattr(cls, attr_name, reset_attrs_decorator(attr_value))
+        elif isinstance(attr_value, property):
+            setattr(
+                cls,
+                attr_name,
+                property(
+                    propagate_attrs_decorator(
+                        attr_value.fget
+                        if attr_value.fget is not None
+                        else attr_value.__get__
+                    ),
+                    propagate_attrs_decorator(
+                        attr_value.fset
+                        if attr_value.fset is not None
+                        else attr_value.__set__
+                    ),
+                    propagate_attrs_decorator(
+                        attr_value.fdel
+                        if attr_value.fdel is not None
+                        else attr_value.__delete__
+                    ),
+                ),
+            )
+        elif inspect.isfunction(attr_value):
+            setattr(cls, attr_name, propagate_attrs_decorator(attr_value))
+    return cls
+
+
+@_propagate_attrs_on_methods
 class SnowflakeQueryCompiler(BaseQueryCompiler):
     """based on: https://modin.readthedocs.io/en/0.11.0/flow/modin/backends/base/query_compiler.html
     this class is best explained by looking at https://github.com/modin-project/modin/blob/a8be482e644519f2823668210cec5cf1564deb7e/modin/experimental/core/storage_formats/hdk/query_compiler.py
@@ -429,6 +518,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         # self.snowpark_pandas_api_calls a list of lazy Snowpark pandas telemetry api calls
         # Copying and modifying self.snowpark_pandas_api_calls is taken care of in telemetry decorators
         self.snowpark_pandas_api_calls: list = []
+        self._attrs: dict[Any, Any] = {}
 
     def _raise_not_implemented_error_for_timedelta(
         self, frame: InternalFrame = None
@@ -854,7 +944,10 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             The QueryCompiler converted to pandas.
 
         """
-        return self._modin_frame.to_pandas(statement_params, **kwargs)
+        result = self._modin_frame.to_pandas(statement_params, **kwargs)
+        if self._attrs:
+            result.attrs = self._attrs
+        return result
 
     def finalize(self) -> None:
         pass
@@ -1607,8 +1700,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         frame = self._modin_frame.ensure_row_position_column()
         row_position_quoted_identifier = frame.row_position_snowflake_quoted_identifier
 
-        fill_value_dtype = infer_object_type(fill_value)
-        fill_value = None if pd.isna(fill_value) else pandas_lit(fill_value)
+        timedelta_invalid_fill_value_error_message = f"value should be a 'Timedelta' or 'NaT'. Got '{type(fill_value).__name__}' instead."
 
         def shift_expression_and_type(
             quoted_identifier: str, dtype: DataType
@@ -1623,19 +1715,51 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             Returns:
                 SnowparkPandasColumn representing the result.
             """
+            if isinstance(dtype, TimedeltaType):
+                if isinstance(fill_value, str):
+                    # Despite the error messages, pandas allows filling a timedelta
+                    # with strings, but it converts strings to timedelta.
+                    try:
+                        fill_value_for_snowpark = pd.Timedelta(fill_value)
+                    except BaseException:
+                        raise TypeError(timedelta_invalid_fill_value_error_message)
+                else:
+                    fill_value_for_snowpark = fill_value
+                if not (
+                    pd.isna(fill_value_for_snowpark)
+                    or isinstance(
+                        SnowparkPandasType.get_snowpark_pandas_type_for_pandas_type(
+                            type(fill_value_for_snowpark)
+                        ),
+                        TimedeltaType,
+                    )
+                ):
+                    raise TypeError(timedelta_invalid_fill_value_error_message)
+            else:
+                fill_value_for_snowpark = fill_value
+
+            fill_value_dtype = infer_object_type(fill_value_for_snowpark)
+            fill_value_snowpark_column = (
+                None
+                if pd.isna(fill_value_for_snowpark)
+                else pandas_lit(fill_value_for_snowpark)
+            )
+
             window_expr = Window.orderBy(col(row_position_quoted_identifier))
 
             # convert to variant type if types differ
-            if fill_value is not None and dtype != fill_value_dtype:
+            if fill_value_snowpark_column is not None and dtype != fill_value_dtype:
                 shift_expression = lag(
                     to_variant(col(quoted_identifier)),
                     offset=periods,
-                    default_value=to_variant(fill_value),
+                    default_value=to_variant(fill_value_snowpark_column),
                 ).over(window_expr)
                 expression_type = VariantType()
             else:
                 shift_expression = lag(
-                    quoted_identifier, offset=periods, default_value=fill_value
+                    quoted_identifier,
+                    offset=periods,
+                    default_value=fill_value_snowpark_column,
                 ).over(window_expr)
                 expression_type = dtype
             # TODO(https://snowflakecomputing.atlassian.net/browse/SNOW-1634393):
@@ -1681,10 +1805,17 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         frame = self._modin_frame
         column_labels = frame.data_column_pandas_labels
 
+        fill_value_snowpark_pandas_type = (
+            SnowparkPandasType.get_snowpark_pandas_type_for_pandas_type(
+                type(fill_value)
+            )
+        )
+
         # Fill all columns with fill value (or NULL) if abs(periods) exceeds column count.
         if abs(periods) >= len(column_labels):
             new_frame = frame.apply_snowpark_function_to_columns(
-                lambda column: pandas_lit(fill_value)
+                lambda column: pandas_lit(fill_value),
+                return_type=fill_value_snowpark_pandas_type,
             )
             return self.__constructor__(new_frame)
 
@@ -1700,18 +1831,25 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             col(quoted_identifier)
             for quoted_identifier in frame.data_column_snowflake_quoted_identifiers
         ]
+        col_snowpark_pandas_types = frame.cached_data_column_snowpark_pandas_types
         if periods > 0:
             # create expressions to shift data to right
             # | lit(...) | lit(...) | ... | lit(...) | col(...) | ... | col(...) |
             col_expressions = [pandas_lit(fill_value)] * periods + col_expressions[
                 :-periods
             ]
+            snowpark_pandas_types = [
+                fill_value_snowpark_pandas_type
+            ] * periods + col_snowpark_pandas_types[:-periods]
         else:
             # create expressions to shift data to left
             # | col(...) | ... | col(...) | lit(...) | lit(...) | ... | lit(...) |
             col_expressions = col_expressions[-periods:] + [pandas_lit(fill_value)] * (
                 -periods
             )
+            snowpark_pandas_types = col_snowpark_pandas_types[-periods:] + [
+                fill_value_snowpark_pandas_type
+            ] * (-periods)
 
         new_frame = frame.update_snowflake_quoted_identifiers_with_expressions(
             {
@@ -1719,7 +1857,8 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
                 for i, quoted_identifier in enumerate(
                     frame.data_column_snowflake_quoted_identifiers
                 )
-            }
+            },
+            snowpark_pandas_types=snowpark_pandas_types,
         ).frame
 
         return self.__constructor__(new_frame)
@@ -6019,6 +6158,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             )
 
         query_compiler = self
+        initial_attrs = self._attrs
         if numeric_only:
             # drop off the non-numeric data columns if the data column if numeric_only is configured to be True
             query_compiler = drop_non_numeric_data_columns(
@@ -6435,6 +6575,11 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             result = result.transpose_single_row()
             # Set the single column's name to MODIN_UNNAMED_SERIES_LABEL
             result = result.set_columns([MODIN_UNNAMED_SERIES_LABEL])
+        # native pandas clears attrs if the aggregation was a list, but propagates it otherwise
+        if is_list_like(func):
+            result._attrs = {}
+        else:
+            result._attrs = copy.deepcopy(initial_attrs)
         return result
 
     def insert(
@@ -7290,6 +7435,10 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
                         raise ValueError(
                             f"Indexes have overlapping values. Few of them are: {overlap}. Please run df1.index.intersection(df2.index) to see complete list"
                         )
+        # If each input's `attrs` was identical and not empty, then copy it to the output.
+        # Otherwise, leave `attrs` empty.
+        if len(self._attrs) > 0 and all(self._attrs == o._attrs for o in other):
+            qc._attrs = copy.deepcopy(self._attrs)
         return qc
 
     def cumsum(
@@ -8317,7 +8466,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
                 ErrorMessage.not_implemented(
                     f"Snowpark pandas apply API doesn't yet support Snowpark Python function `{func.__name__}` with args = '{args}'."
                 )
-            return self._apply_snowpark_python_function_to_columns(func)
+            return self._apply_snowpark_python_function_to_columns(func, kwargs)
 
         if axis == 0:
             frame = self._modin_frame
@@ -8576,11 +8725,37 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
     def _apply_snowpark_python_function_to_columns(
         self,
         snowpark_function: Callable,
+        kwargs: dict[str, Any],  # possible named arguments which need to be added
     ) -> "SnowflakeQueryCompiler":
         """Apply Snowpark Python function to columns."""
 
         def sf_function(col: SnowparkColumn) -> SnowparkColumn:
-            return snowpark_function(col)
+            if not kwargs:
+                return snowpark_function(col)
+            # we have named kwargs, which may be positional
+            # in nature, and we need to align them to the snowpark
+            # function call alongside the column reference
+            # Get the total arg count for the function
+            function_arg_count = snowpark_function.__code__.co_argcount
+            # Get all variables for the function and slice off only the arguments
+            positional_args = snowpark_function.__code__.co_varnames[
+                :function_arg_count
+            ]
+            resolved_positional = []
+            col_specified = False
+            for arg in positional_args:
+                if arg in kwargs:
+                    resolved_positional.append(kwargs[arg])
+                else:
+                    if not col_specified:
+                        resolved_positional.append(col)
+                        col_specified = True
+                    else:
+                        ErrorMessage.not_implemented(
+                            f"Unspecified Argument: {arg} - when using apply with kwargs, all function arguments should be specified except the single column reference (if applicable)."
+                        )
+
+            return snowpark_function(*resolved_positional)
 
         return SnowflakeQueryCompiler(
             self._modin_frame.apply_snowpark_function_to_columns(sf_function)
@@ -8615,7 +8790,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
                 ErrorMessage.not_implemented(
                     f"Snowpark pandas applymap API doesn't yet support Snowpark Python function `{func.__name__}` with args = '{args}'."
                 )
-            return self._apply_snowpark_python_function_to_columns(func)
+            return self._apply_snowpark_python_function_to_columns(func, kwargs)
         # Currently, NULL values are always passed into the udtf even if strict=True,
         # which is a bug on the server side SNOW-880105.
         # The fix will not land soon, so we are going to raise not implemented error for now.
@@ -8654,6 +8829,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         self,
         arg: Union[AggFuncType, "pd.Series"],
         na_action: Optional[Literal["ignore"]] = None,
+        **kwargs: Any,
     ) -> "SnowflakeQueryCompiler":
         """This method will only be called from Series."""
         self._raise_not_implemented_error_for_timedelta()
@@ -8671,7 +8847,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             ErrorMessage.not_implemented(
                 "Snowpark pandas map API doesn't yet support non callable 'arg'"
             )
-        return self.applymap(func=arg, na_action=na_action)
+        return self.applymap(func=arg, na_action=na_action, **kwargs)
 
     def apply_on_series(
         self, func: AggFuncType, args: tuple[Any, ...] = (), **kwargs: Any
@@ -11852,9 +12028,13 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         return self._modin_frame.is_multiindex(axis=axis)
 
     def abs(self) -> "SnowflakeQueryCompiler":
+        # TODO(SNOW-1620415): Implement abs() for timedelta.
+        self._raise_not_implemented_error_for_timedelta()
         return self.unary_op("abs")
 
     def negative(self) -> "SnowflakeQueryCompiler":
+        # TODO(SNOW-1620415): Implement __neg__() for timedelta.
+        self._raise_not_implemented_error_for_timedelta()
         return self.unary_op("__neg__")
 
     def unary_op(self, op: str) -> "SnowflakeQueryCompiler":
@@ -12633,8 +12813,6 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
                 rather than the entire dataset. This parameter is exclusive to the Snowpark pandas
                 query compiler and is only used internally to implement groupby_value_counts.
         """
-        self._raise_not_implemented_error_for_timedelta()
-
         # validate whether by is valid (e.g., contains duplicates or non-existing labels)
         self.validate_groupby(by=by, axis=0, level=None)
 
@@ -18225,6 +18403,11 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         sort : bool, default True
             Whether to sort the levels of the resulting MultiIndex.
         """
+        # stack() may create a column that includes values from multiple input
+        # columns. Tracking types in that case is not simple, so we don't
+        # handle the client-side timedelta type as an input.
+        self._raise_not_implemented_error_for_timedelta()
+
         if level != -1:
             ErrorMessage.not_implemented(
                 "Snowpark pandas doesn't yet support 'level != -1' in stack API",
@@ -18284,6 +18467,11 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             ErrorMessage.not_implemented(
                 "Snowpark pandas doesn't support multiindex columns in the unstack API"
             )
+
+        # unstack() should preserve timedelta types, but one input column may
+        # may map to multiple output columns, so we don't support timedelta
+        # inputs yet.
+        self._raise_not_implemented_error_for_timedelta()
 
         level = [level]